src/google_cloud/firestore_events/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
backoff==2.2.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
deprecation==2.1.0
distro==1.8.0
Flask==2.3.3
functions-framework==3.4.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-secret-manager==2.16.4
google-cloud-tasks==2.15.0
google-events==0.11.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.7
grpcio==1.59.2
grpcio-status==1.59.2
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
openai==1.3.5
packaging==23.2
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
regex==2023.10.3
requests==2.31.0
rsa==4.9
setuptools==68.2.2
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
watchdog==3.0.0
Werkzeug==3.0.1
wheel==0.41.2
